changed the folder structure so there is less api
