--- restapi/api/api.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.3.32811.315
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "api", "api\api.csproj", "{75CB1F3E-C1B0-44CA-827E-65CBBFC14E2E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{75CB1F3E-C1B0-44CA-827E-65CBBFC14E2E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{75CB1F3E-C1B0-44CA-827E-65CBBFC14E2E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{75CB1F3E-C1B0-44CA-827E-65CBBFC14E2E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{75CB1F3E-C1B0-44CA-827E-65CBBFC14E2E}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {33551C96-AFC8-4E39-A29D-E44401B23FE9}
-	EndGlobalSection
-EndGlobal
